[TASK] bamboo: use public DNS name, not server name
This makes it easier to migrate to another server in case we need to.

The old server name is specific to the machine whereas the new name does
not change when we move to another server i.e. the DNS entry would
simply point to anther IP and the build config does not need to be
updated.

Related: INFRA-160
Releases: master