[BUGFIX] bamboo: move unneeded config files into 'backup' dir
These files were downloaded as a temporary 'backup' to have a safety net
before changing the jobs.

Unfortunately these files breat the autoloader when building the
project which blocks the deployment of build spec changes:

    $ mvn test
    ...
    [INFO] BUILD FAILURE
    [INFO] ------------------------------------------------------------------------
    [INFO] Total time:  1.549 s
    [INFO] Finished at: 2020-10-16T12:04:40+02:00
    [INFO] ------------------------------------------------------------------------
    [ERROR] Failed to execute goal org.apache.maven.plugins:maven-compiler-plugin:3.7.0:compile (default-compile) on project typo3-intercept: Compilation failure: Compilation failure:
    [ERROR] build/bamboo/src/main/java/docs/DocsRenderingApi.java:[34,8] class PlanSpec is public, should be declared in a file named PlanSpec.java
    [ERROR] build/bamboo/src/main/java/docs/DocsRenderingApiRoot.java:[30,8] class PlanSpec is public, should be declared in a file named PlanSpec.java

This commit ensures that the backup files can be used as reference and
ensures that they don't interfere with the build itself so changes to
existing builds can be deployed.

Fixup for 048ea0ccadf77729ce01b17fed830df4b1e1005d

Ticket(s): INFRA-160